Moved Setting code!
